+++ CHANGELOG.md
@@ -3,7 +3,8 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 ### Fixed
-- Transaction max value validation
+- Transaction max value validation #994
+- Checking signature hash correctness #995
 
 ## [v2.1.4] - 2020-03-27
 ### Changed